Start reimplementing alternative cam modes
